+++ internal/store/cache/mocks.go
@@ -17,7 +17,7 @@ type MockUserStore struct {
 	mock.Mock
 }
 
-func (m *MockUserStore) Get(ctx context.Context, userID int64) (*store.User, error) {
+func (m *MockUserStore) Get(ctx context.Context, userID uint) (*store.User, error) {
 	args := m.Called(userID)
 	return nil, args.Error(1)
 }
@@ -27,6 +27,6 @@ func (m *MockUserStore) Set(ctx context.Context, user *store.User) error {
 	return args.Error(0)
 }
 
-func (m *MockUserStore) Delete(ctx context.Context, userID int64) {
+func (m *MockUserStore) Delete(ctx context.Context, userID uint) {
 	m.Called(userID)
 }
